Get all tests to pass after embedding InitResolver in Resolver
