+++ frontend/lib/resolution/Makefile.include
@@ -25,7 +25,7 @@ FRONTEND_RESOLUTION_SRCS =                          \
            can-pass.cpp                             \
            default-functions.cpp                    \
            disambiguation.cpp                       \
-           init-resolver-visitor.cpp                \
+           InitResolver.cpp                         \
            intents.cpp                              \
            prims.cpp                                \
            resolution-queries.cpp                   \
